Started on creating classes in for use_cases in order to make our UML diagram and sequence. Organized the src files for our actual app.
