+++ src/use_case/CalculateTerminalValue/CalculateTerminalValue.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.CalculateTerminalValue;
 
 public class CalculateTerminalValue {
 }
