+++ src/use_case/CalculateTerminalValue/CalculateTerminalValueInputBoundary.java
@@ -1,0 +1,4 @@
+package use_case.CalculateTerminalValue;
+
+public interface CalculateTerminalValueInputBoundary {
+}
