+++ src/use_case/CalculateTerminalValue/CalculateTerminalValueInputData.java
@@ -1,0 +1,4 @@
+package use_case.CalculateTerminalValue;
+
+public class CalculateTerminalValueInputData {
+}
