+++ src/use_case/CalculateTerminalValue/CalculateTerminalValueOutputBoundary.java
@@ -1,0 +1,4 @@
+package use_case.CalculateTerminalValue;
+
+public interface CalculateTerminalValueOutputBoundary {
+}
