+++ src/use_case/GetDCF/GetDCF.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.GetDCF;
 
 public class GetDCF {
 }
