+++ src/use_case/GetDCF/getDCFInputBoundary.java
@@ -1,0 +1,4 @@
+package use_case.GetDCF;
+
+public interface getDCFInputBoundary {
+}
