+++ src/use_case/GetDCF/getDCFInputData.java
@@ -1,0 +1,4 @@
+package use_case.GetDCF;
+
+public class getDCFInputData {
+}
