+++ src/use_case/GetDCF/getDCFOutputData.java
@@ -1,0 +1,4 @@
+package use_case.GetDCF;
+
+public class getDCFOutputData {
+}
