+++ src/use_case/GetDCF/getDCOutputBoundary.java
@@ -1,0 +1,4 @@
+package use_case.GetDCF;
+
+public interface getDCOutputBoundary {
+}
